feat: implement WebAppFrameBridgeService to manage iframe messaging, schema validation, and data synchronization
